Moves all domain SimpleTests to Functional tests.
