+++ domain/tests/src/Functional/DomainCSSTest.php
@@ -1,8 +1,9 @@
 <?php
 
-namespace Drupal\domain\Tests;
+namespace Drupal\Tests\domain\Functional;
 
 use Drupal\Component\Utility\Html;
+use Drupal\Tests\domain\Functional\DomainTestBase;
 
 /**
  * Tests the domain CSS configuration.
